revert back to one file with multiple services
