+++ service/run
@@ -1,7 +1,5 @@
 #!/bin/sh
 SCRIPT_DIR=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )
-NEWPATH_INV="$(dirname "$SCRIPT_DIR")/dbus-huaweisun2000-pvinverter.py"
-NEWPATH_METER="$(dirname "$SCRIPT_DIR")/dbus-huaweisun2000-meter.py"
+NEWPATH="$(dirname "$SCRIPT_DIR")/dbus-huaweisun2000-pvinverter.py"
 exec 2>&1
-python -u $NEWPATH_INV
-python -u $NEWPATH_METER
+python -u $NEWPATH
